move proxy classes into poxy package
